Use older pax:runner version, new one won't start

+++ alitheia/core/src/main/java/eu/sqooss/impl/service/updater/OhlohUpdater.java
@@ -75,7 +75,7 @@ public class OhlohUpdater extends UpdaterBaseJob {
     }
 
     @Override
-    public int priority() {
+    public long priority() {
         return 3;
     }
